+++ src/O4_Tile_Utils.py
@@ -38,6 +38,8 @@ def download_textures(tile,download_queue,convert_queue):
 
 ##############################################################################
 def build_tile(tile):
+    if UI.is_working: return 0
+    UI.is_working=1
     UI.red_flag=False
     UI.logprint("Step 3 for tile lat=",tile.lat,", lon=",tile.lon,": starting.")
     UI.vprint(0,"\nStep 3 : Building DSF/Imagery for tile "+FNAMES.short_latlon(tile.lat,tile.lon)+" : \n--------\n")
@@ -59,8 +61,9 @@ def build_tile(tile):
             os.makedirs(os.path.join(tile.build_dir,'Earth nav data',FNAMES.round_latlon(tile.lat,tile.lon)))
         if not os.path.isdir(os.path.join(tile.build_dir,'textures')):
             os.makedirs(os.path.join(tile.build_dir,'textures'))
-        try: shutil.rmtree(os.path.join(tile.build_dir,'terrain'))
-        except: pass
+        if not tile.grouped:    
+            try: shutil.rmtree(os.path.join(tile.build_dir,'terrain'))
+            except: pass
         if not os.path.isdir(os.path.join(tile.build_dir,'terrain')):
             os.makedirs(os.path.join(tile.build_dir,'terrain'))
     except Exception as e: 
@@ -100,12 +103,20 @@ def build_tile(tile):
 ##############################################################################
 def build_all(tile):
     VMAP.build_poly_file(tile)
+    if UI.red_flag: UI.exit_message_and_bottom_line(''); return 0
     MESH.build_mesh(tile)
+    if UI.red_flag: UI.exit_message_and_bottom_line(''); return 0
     MASK.build_masks(tile)
+    if UI.red_flag: UI.exit_message_and_bottom_line(''); return 0
     build_tile(tile)
+    if UI.red_flag: UI.exit_message_and_bottom_line(''); return 0
+    UI.is_working=0
+    return 1
 ##############################################################################
 
 def build_tile_list(tile,list_lat_lon,do_osm,do_mesh,do_mask,do_dsf,do_ovl,do_ptc):
+    if UI.is_working: return 0
+    UI.is_working=1
     timer=time.time()
     UI.lvprint(0,"Batch build launched for a number of",len(list_lat_lon),"tiles.")
     k=0
@@ -128,5 +139,6 @@ def build_tile_list(tile,list_lat_lon,do_osm,do_mesh,do_mask,do_dsf,do_ovl,do_pt
         if do_ovl: OVL.build_overlay(lat,lon)
         if UI.red_flag: UI.exit_message_and_bottom_line(); return 0
     UI.lvprint(0,"Batch process completed in",UI.nicer_timer(time.time()-timer))
+    UI.is_working=0
     return 1
